First working pass at migrating the concurrent chunks from DmrppArray into SuperChunk.
