+++ modules/dmrpp_module/DmrppArray.h
@@ -214,6 +214,7 @@ void process_chunks_unconstrained_concurrent(
         DmrppArray *array,
         const vector<unsigned int> &array_shape);
 
+bool get_next_future(list<std::future<bool>> &futures, atomic_uint &thread_counter, unsigned long timeout, string debug_prefix);
 
 } // namespace dmrpp
